Don't bind finishedFunc into a closure at animation start
This is to resolve a hypothetical situation where, after UIFrameFade is
called, code may attempt to modify the '.finishedFunc' or any argument
fields on the fadeInfo table.

The existing code would, when starting the animation, generate a bound
closure with the func and its arguments fixed. When the animation finished
this closure would be invoked - which if the values on the fadeInfo table
were modified would lead to different behavior compared to the default
implementation of UIFrameFade.

Now, we don't generate a closure and instead just trigger the callback
as-defined on the fadeInfo table at the time when the animation finishes. This
also gives us a small memory usage optimization as we're not having to
generate closures.