Always use frame alpha when starting animation
This removes the copying of the '.fadeInfo' table when initially
importing frames from FADEFRAMES. Instead of using the '.startAlpha'
key we'll unconditionally always use the initial frame alpha.

This change was made since we're not going with any sort of public
API, so we don't need to really support users calling the
StartFadingFrame function.