Explicitly only release faders if they're active
This check is actually handled already by the implementation of
the object pool release method, but rather than rely on that to
forevermore be the case we'll explicitly check here.

The reason for needing to check is that the resetter function will
stop the fader animation, which can trigger a second attempted
release of the fader.