Fix ElvUI chat frame tabs incorrectly fading
Fixes #2. ElvUI installs a posthooked SetAlpha method on the chat tabs
which, if a 'skip' parameter is not manually supplied, makes ElvUI
re-call SetAlpha to force the chat tabs to not fade.

This posthook wasn't being triggered by the library due to the use
of animations, so a heuristic has been added to detect frames that
have hooked or replaced SetAlpha methods present. If found, the fader
will, on each game tick, call the SetAlpha method with the current
smoothed progress of the animation.