Implement UIFrameIsFading
Some addons require UIFrameIsFading to actually return correct values,
which it no longer would if the library were loaded as we need to clear
the FADEFRAMES table.

Assuming the taint work present in Classic makes its way to Mainline,
this replacement - and our manipulation of FADEFRAMES with it - can be
removed in the future.

Until then, we need to provide a custom implementation of the
UIFrameIsFading function for a handful of addons to not break.

The UIFrameIsFading function is entirely unused by Blizzard as of the
following current game versions:

    - 9.1.0
    - 2.5.2
    - 1.14.0

While our replacement of this does inherently taint the function
itself, no secure callers should ever be impacted.