Rework handling of secure hooks
This reworks our internal usage of hooksecurefunc to permit the
ability to later unhook functions if our implementation requirements
change.

One caveat is that unhooking will only work if the first loaded
version is v4 or newer, but doing this work now should make that
essentially a guarantee as opposed to doing it weeks (or months)
down the line.

Should special logic be needed to handle pre-v4 hooks, the existing
booleans for their state are kept on the library table.