Disable library in non-mainline clients
The fundamental issue with UIFrameFade taint was fixed long ago
in Classic clients, so this library shouldn't be necessary in those
any more. Would rather not keep it alive longer than it needs to be,
especially given the complicated issue with the combo frame points.

For mainline the issue still exists, but will need to evaluate how
problematic any bugs actually are there - especially as UIFrameFade
is in much smaller use in modern UI code.