Update for modern client versions
The library is no longer required on any modern version of Classic or
Mainline, and as such can be globally disabled. Just in case the code is
needed again it's all here and intact, but there's now a global toggle
defined that turns the library off as needed.

Additionally; update for modern pools. SetResetDisallowedIfNew was
removed, but was just an optimization anyway.